integrations/requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0
websockets==10.4
selenium==4.11.2
beautifulsoup4==4.12.2
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.0
matplotlib==3.7.2
loguru==0.7.0
# Choose one of these, not both
pyppeteer==1.0.2
# playwright==1.38.0
jsonschema==4.19.0
pymongo==4.5.0
redis==4.6.0
prometheus-client==0.17.1
pydantic>=2.6.1,<3.0.0
diskcache==5.6.3
tiktoken==0.5.2
docker==6.1.3
termcolor==2.4.0
fast-depends==2.4.12
ag2[openai]==0.7.6
aiohttp==3.9.1
